Update the graph reading process in k-coloring puzzle
We will directly read data from adjacency list. We previously changed the constraint graph input format as we were calling the nodes to assign color serially and at each step check whether the color can be assigned in that node. We now choose node to visit either randomly or by heuristic method, and check whether the color assignment is valid or not at the end of assigning color to every nodes. So we don't need to read the nodes first to remove the bias.